+++ src/models/ProxyMachine.php
@@ -2,9 +2,11 @@
 
 namespace luya\admin\models;
 
+use luya\admin\aws\DetailViewActiveWindow;
 use Yii;
 use luya\admin\ngrest\base\NgRestModel;
 use luya\admin\traits\SoftDeleteTrait;
+use luya\helpers\Url;
 
 /**
  * Proxy Machine.
@@ -113,6 +115,31 @@ class ProxyMachine extends NgRestModel
             [['list'], ['name', 'identifier', 'access_token', 'is_disabled']],
             [['create', 'update'], ['name']],
             [['delete'], true],
+        ];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function ngRestActiveWindows()
+    {
+        return [
+            [
+                'class' => DetailViewActiveWindow::class,
+                'intro' => 'For more details about how to use the sync command, take a look at the <a class="text-primary" href="https://luya.io/api/luya-admin-commands-ProxyController" target="_blank">admin/proxy guide</a>.',
+                'attributes' => [
+                    'name',
+                    'identifier',
+                    'access_token',
+                    [
+                        'label' => 'cli',
+                        'format' => 'html',
+                        'value' => function($model) {
+                            return '<code>admin/proxy --url='.Url::base(true).' --idf='.$model->identifier.' --token=' . $model->access_token . '</code>';
+                        }
+                    ]
+                ]
+            ]
         ];
     }
 }