+++ src/proxy/ClientBuild.php
@@ -10,6 +10,8 @@ use yii\db\Connection;
 
 /**
  * Admin Proxy Build.
+ * 
+ * For `admin/proxy` usage see {{luya\admin\commands\ProxyController}}
  *
  * @author Basil Suter <basil@nadar.io>
  * @since 1.0.0